server_config/python/requirements.txt
jsonschema==4.17.3
python-dotenv==1.0.0
pathlib==1.0.1
configparser==5.3.0
mysql-connector-python==8.2.0
requests==2.31.0
Flask==3.0.0
minio==7.1.17
python-telegram-bot==20.7